updated MLKit licences & rename 'VisionKit' to 'Vision'
