+++ source/MLKit/Vision/External-Dependency-Info.txt
