+++ source/MLKit/Vision/Vision.targets
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="UTF-8"?>
 <Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
 	<PropertyGroup>
-		<_MLKitVisionRecogCoreAssemblyName>MLKit.VisionKit, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null</_MLKitVisionRecogCoreAssemblyName>
+		<_MLKitVisionRecogCoreAssemblyName>MLKit.Vision, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null</_MLKitVisionRecogCoreAssemblyName>
 		<_MLKitVisionKitItemsFolder>MLKVisionKit-3.0.0</_MLKitVisionKitItemsFolder>
 		<_MLKitVisionKitSDKBaseFolder>$(XamarinBuildDownloadDir)$(_MLKitVisionKitItemsFolder)\</_MLKitVisionKitSDKBaseFolder>
 		<_MLKitImgLblCommonItemsFolder>MLKImgLblCommon-2.0.0</_MLKitImgLblCommonItemsFolder>
